fix issues

+++ Reviiyot/Game.h
@@ -22,24 +22,24 @@ class Game {
 private:
 	vector<Player *> players;  //The list of the players
 	Deck deck;                 //The deck of the game
-	vector<int> *cardCount;		// Struct to count the player's card amounts.
+	vector<unsigned long> *cardCount;		// Struct to count the player's card amounts.
 	const char* conf;
 
-	unsigned int verbalConfig;
-	unsigned int highestNum;
+	unsigned long verbalConfig;
+	unsigned long highestNum;
 
 public:
 	Game(char* configurationFile);
 	~Game();
 	void init();
 	void play();
-	void printTurn(int turn);
+	void printTurn(unsigned long turn);
 	void printAsk(string p1, string p2, Card & card);
 	void printState();        //Print the state of the game as described in the assignment.
 	void printWinner();       //Print the winner of the game as describe in the assignment.
 	void printNumberOfTurns(); //Print the number of played turns at any given time.  
 	void parseConfig();
-	Player * createPlayer(char type, string name, int pos);
+	Player * createPlayer(char type, string name, unsigned long pos);
 };
 
 #endif
