+++ Reviiyot/Player.cpp
@@ -1,6 +1,6 @@
 #include "Player.h"
 
-Player::Player(string name, int pos):Hand(), pos(pos),name(name)
+Player::Player(string name, unsigned long pos):Hand(), pos(pos),name(name)
 {
 	cout << "player created" << endl;
 }
@@ -21,7 +21,7 @@ string Player::toString()
 	return name + ": " + Hand::toString();
 }
 
-PlayerType1::PlayerType1(string name, int pos):Player(name,pos)
+PlayerType1::PlayerType1(string name, unsigned long pos):Player(name,pos)
 {
 	cout << "player 1 created" << endl;
 
@@ -32,7 +32,7 @@ PlayerType1::~PlayerType1()
 	cout << "player 1 deleted" << endl;
 }
 
-pair<int, Card&> PlayerType1::ask(vector<int> state)
+pair<unsigned long, Card&> PlayerType1::ask(vector<unsigned long> state)
 {
 	// Descover who
 	int who = 0;
@@ -55,7 +55,7 @@ pair<int, Card&> PlayerType1::ask(vector<int> state)
 	return pair<int, Card&>(who, what.second);
 }
 
-PlayerType2::PlayerType2(string name, int pos) :Player(name, pos)
+PlayerType2::PlayerType2(string name, unsigned long pos) :Player(name, pos)
 {
 	cout << "player 2 created" << endl;
 }
@@ -65,10 +65,10 @@ PlayerType2::~PlayerType2()
 	cout << "player 2 deleted" << endl;
 }
 
-pair<int, Card&> PlayerType2::ask(vector<int> state)
+pair<unsigned long, Card&> PlayerType2::ask(vector<unsigned long> state)
 {
 	// Descover who
-	int who = 0;
+	unsigned long who = 0;
 	if (pos == 0)
 	{
 		who = 1;
@@ -85,10 +85,10 @@ pair<int, Card&> PlayerType2::ask(vector<int> state)
 	// Descover what
 	pair<int, Card&> what = getTheLeast();
 
-	return pair<int, Card&>(who, what.second);
+	return pair<unsigned long, Card&>(who, what.second);
 }
 
-PlayerType3::PlayerType3(string name, int pos) :Player(name, pos)
+PlayerType3::PlayerType3(string name, unsigned long pos) :Player(name, pos)
 {
 	cout << "player 3 created" << endl;
 
@@ -100,10 +100,10 @@ PlayerType3::~PlayerType3()
 	cout << "player 3 deleted" << endl;
 }
 
-pair<int, Card&> PlayerType3::ask(vector<int> state)
+pair<unsigned long, Card&> PlayerType3::ask(vector<unsigned long> state)
 {
 	// Descover who
-	int who = nextAsk;
+	unsigned long who = nextAsk;
 	if (pos == who)
 	{
 		who = (who + 1) % state.size();
@@ -114,10 +114,10 @@ pair<int, Card&> PlayerType3::ask(vector<int> state)
 	// Descover what
 	Card& what = *getHighestCard();
 
-	return pair<int, Card&>(who, what);
+	return pair<unsigned long, Card&>(who, what);
 }
 
-PlayerType4::PlayerType4(string name, int pos):Player(name,pos)
+PlayerType4::PlayerType4(string name, unsigned long pos):Player(name,pos)
 {
 	cout << "player 4 created" << endl;
 
@@ -129,10 +129,10 @@ PlayerType4::~PlayerType4()
 	cout << "player 4 deleted" << endl;
 }
 
-pair<int, Card&> PlayerType4::ask(vector<int> state)
+pair<unsigned long, Card&> PlayerType4::ask(vector<unsigned long> state)
 {
 	// Descover who
-	int who = nextAsk;
+	unsigned long who = nextAsk;
 	if (pos == who)
 	{
 		who = (who + 1) % state.size();
@@ -143,5 +143,5 @@ pair<int, Card&> PlayerType4::ask(vector<int> state)
 	// Descover what
 	Card& what = *getLowestCard();
 
-	return pair<int, Card&>(who, what);
+	return pair<unsigned long, Card&>(who, what);
 }
